On branch master:
Create setup.sh and setup.py to configure the linux environment.
setup.sh update the local packege database, then checks if python3 is installed.
Once python3 is ready, the script will start running setup.py
setup.py installs and configures openssh-server and xrdp, then creates a user for remote desktop access.
The setting is interactive. You have to be careful when running it, because it expects input from the user several times.
After a successful run, the machine can be accessed via ssh and RDP.

Changes to be committed:
	new file:   setup.py
	new file:   setup.sh